Moved getFragmentName and favourite option to FragmentManagement.
Also, renamed FragmentManager to FragmentManagement to avoid conflicts.
