--- app/src/main/java/swati4star/createpdf/providers/FragmentManager.java
@@ -1,9 +1,0 @@
-package swati4star.createpdf.providers;
-
-/**
- * This is a fragment service that manages the fragments
- * mainly for the MainActivity.
- */
-public class FragmentManager implements IFragmentManager {
-
-}
