+++ app/src/main/java/swati4star/createpdf/providers/IFragmentManagement.java
@@ -1,0 +1,8 @@
+package swati4star.createpdf.providers;
+
+interface IFragmentManagement {
+    /**
+     * Begins a fragment transaction for the favourite fragment.
+     */
+    void favouritesFragmentOption();
+}
