--- app/src/main/java/swati4star/createpdf/providers/IFragmentManager.java
@@ -1,4 +1,0 @@
-package swati4star.createpdf.providers;
-
-public interface IFragmentManager {
-}
